backend/requirements.txt
flask
flask-jwt-extended
bcrypt
mysql-connector-python
